Directives
are instuctions in the DOM(document object modle)...
1)ngif  ($ngif="value")
2)ngif="value; else asd)
ngtemplate #asd (inline if else html angular)

attribute directivise doesnt add or remove elements. they only change elments they are placed on.

structural directives are the ones which start with a *, that do add or remove to/from the DOM

3) [ngStyle] the square brackets indicates that its being binded to a property

4) ngClass "  [ngClass]="{ online: serverStatus === 'Online' }"
" if serverStatus value is online then apply that class.

5)*ngFor="let server of servers" loop on servers array and out put the data

then practice on an old practice

note: # is a local refrence